added middleware, database handling

+++ handlers/handlers.go
@@ -6,6 +6,7 @@ import (
 	"github.com/gofiber/fiber/v2"
 	"github.com/sourabh-sharma3435/rest-inventory/models"
 	"github.com/sourabh-sharma3435/rest-inventory/services"
+	"github.com/sourabh-sharma3435/rest-inventory/utils"
 )
 
 func GetAllItems(c *fiber.Ctx) error {
@@ -37,6 +38,14 @@ func GetItemByID(c *fiber.Ctx) error {
 }
 
 func CreateItem(c *fiber.Ctx) error {
+	isvalid, err := utils.CheckToken(c)
+	if !isvalid {
+		return c.Status(http.StatusUnauthorized).JSON(models.Response[any]{
+			Success: false,
+			Message: err.Error(),
+		})
+	}
+
 	var itemInput *models.ItemRequest = new(models.ItemRequest)
 	if err := c.BodyParser(itemInput); err != nil {
 		return c.Status(http.StatusBadRequest).JSON(models.Response[any]{
@@ -65,6 +74,15 @@ func CreateItem(c *fiber.Ctx) error {
 }
 
 func UpdateItem(c *fiber.Ctx) error {
+	isValid, err := utils.CheckToken(c)
+
+	if !isValid {
+		return c.Status(http.StatusUnauthorized).JSON(models.Response[any]{
+			Success: false,
+			Message: err.Error(),
+		})
+	}
+
 	var itemInput *models.ItemRequest = new(models.ItemRequest)
 
 	if err := c.BodyParser(itemInput); err != nil {
@@ -101,6 +119,15 @@ func UpdateItem(c *fiber.Ctx) error {
 }
 
 func DeleteItem(c *fiber.Ctx) error {
+	isValid, err := utils.CheckToken(c)
+
+	if !isValid {
+		return c.Status(http.StatusUnauthorized).JSON(models.Response[any]{
+			Success: false,
+			Message: err.Error(),
+		})
+	}
+
 	var itemID string = c.Params("id")
 
 	var result bool = services.DeleteItem(itemID)
